git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@7298 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/ports/STM32/LLD/TIMv1/icu_lld.c
@@ -508,7 +508,7 @@ void icu_lld_start(ICUDriver *icup) {
   osalDbgAssert((psc <= 0xFFFF) &&
                 ((psc + 1) * icup->config->frequency) == icup->clock,
                 "invalid frequency");
-  icup->tim->PSC  = (uint16_t)psc;
+  icup->tim->PSC  = psc;
   icup->tim->ARR  = 0xFFFF;
 
   if (icup->config->channel == ICU_CHANNEL_1) {
