+++ os/hal/ports/STM32/LLD/TIMv1/icu_lld.h
@@ -256,7 +256,7 @@ typedef enum {
 /**
  * @brief   ICU counter type.
  */
-typedef uint16_t icucnt_t;
+typedef uint32_t icucnt_t;
 
 /**
  * @brief   Driver configuration structure.
